Inserted SERIOUS HACK into patch visibility function. Added day-aware ambient illumination. Fixed tracked navigation orientation. Reworked near/far clipping to use radius instead of Z. Changed controller joystick to update even when not changing. Moved most tweakables to default.h. Added joystick rotation of earth. Fixed atmosphere not rendered when earth is out of view.
git-svn-id: https://s.snth.net/svn/rlk/thumb@1071 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ src/universe.hpp
@@ -39,7 +39,7 @@ namespace uni
         void draw(const double *, const double *);
 
         double rate() const;
-        void turn(int);
+        void turn(double);
     };
 }
